@@ -9,12 +9,8 @@ from aiohttp import ClientResponse
 from aiohttp.http_websocket import WebSocketError
 from asyncprawcore import BadRequest, Forbidden, NotFound, TooLarge
 
-if sys.version_info < (3, 8):
-    from asynctest import CoroutineMock as AsyncMock
-    from asynctest import MagicMock, mock
-else:
-    from unittest import mock
-    from unittest.mock import AsyncMock, MagicMock
+from unittest import mock
+from unittest.mock import AsyncMock, MagicMock
 
 from asyncpraw.const import PNG_HEADER
 from asyncpraw.exceptions import (
