@@ -4,10 +4,7 @@ from pathlib import Path
 
 import pytest
 
-if sys.version_info < (3, 8):
-    from asynctest import mock
-else:
-    from unittest import mock
+from unittest import mock
 
 from asyncpraw.config import Config
 from asyncpraw.exceptions import ClientException
