@@ -5,10 +5,7 @@ from tempfile import NamedTemporaryFile
 import aiofiles
 import pytest
 
-if sys.version_info < (3, 8):
-    from asynctest import mock
-else:
-    from unittest import mock
+from unittest import mock
 
 from asyncpraw.util.token_manager import (
     BaseTokenManager,
